Update and rename .gitignore to .gitignore102CEM
Assignment Brief: Pong is one of the earlierst video games ever sold. The idea is very simple: two players each control a "bat" or "paddle" on an edge of the screen. They can move the bat back and forth in order to bounce a ball that is travelling across the screen. If they miss and the ball hits their edge of the screen, the opponent gains a point. 1. Build a controller for a single player 2. Write the code to use that interface to control one of the bats/paddles 3. Add a second controller - you can work with a partner for this second paddle. You gain marks only for the code that makes use of it, not the controller itself 4. Extend the game to include extra features such as: (a) Random ball starting location/direction • I suggest keeping the random locations within the middle 6 LEDs to avoid hitting an edge too soon • Also the ball really only works well going diagonally, so the direction must be one of (-1,1),(-1,-1),(1,-1),(1,1). 2 (b) A reset/restart button • Something mroe than just the reset of the Arduino. Perhaps you can do a fancy screen wipe of the matrix? A loop that fills each line in red then blanks them again. . . (c) A better score display than the serial output • Many options here. There are 7-segment LED units in the kit, but you could also just line up two lots of 3 LEDs. Maybe 2 blue and 1 red for each player, so that when you get the red one lit, you win. . . ? Or maybe you could use processing to show the score ont he PC. Use your imagination. (d) Halting the game when someone reaches a maximum score (say, 3) and is therefore the winner. • Even better if you can show a message or icon on the LED matrix to declare the winner